renamed interface to ontology; moved owl-file; worked on owl-file
