--- probeye/interface/__init__.py
@@ -1,2 +1,0 @@
-# local imports
-from probeye.interface import knowledge_graph_export
