+++ setup.cfg
@@ -37,7 +37,7 @@ install_requires =
 [options.package_data]
 probeye = 
     probeye.txt
-    parameter_estimation_ontology.owl
+    ontology/parameter_estimation_ontology.owl
 
 [options.extras_require]
 tests =
