+++ tests/integration_tests/tests_with_correlation/test_1D_1D_correlation_additive_model_error_1std.py
@@ -32,7 +32,7 @@ from probeye.definition.sensor import Sensor
 from probeye.subroutines import len_or_one
 
 # local imports (knowledge graph)
-from probeye.interface.knowledge_graph_export import export_knowledge_graph
+from probeye.ontology.knowledge_graph_export import export_knowledge_graph
 
 # local imports (testing related)
 from tests.integration_tests.subroutines import run_inference_engines
